+++ src/app/Http/Controllers/Operations/UpdateOperation.php
@@ -95,7 +95,7 @@ trait UpdateOperation
 
         // update the row in the db
         $item = $this->crud->update($request->get($this->crud->model->getKeyName()),
-                            $request->except('save_action', '_token', '_method', 'current_tab', 'http_referrer'));
+                            $this->crud->getStrippedSaveRequest());
         $this->data['entry'] = $this->crud->entry = $item;
 
         // show a success message