+++ src/app/Library/CrudPanel/Traits/FakeFields.php
@@ -25,14 +25,19 @@ trait FakeFields
         $compactedFakeFields = [];
         foreach ($fields as $field) {
             // compact fake fields
-            if (isset($field['fake']) && $field['fake'] == true && array_key_exists($field['name'], $requestInput)) {
-                $fakeFieldKey = isset($field['store_in']) ? $field['store_in'] : 'extras';
-                $this->addCompactedField($requestInput, $field['name'], $fakeFieldKey);
+            // cast the field name to array first, to account for array field names
+            // in fields that send multiple inputs and want them all saved to the database 
+            foreach ((array)$field['name'] as $fieldName) {
+                if (isset($field['fake']) && $field['fake'] == true && array_key_exists($fieldName, $requestInput)) {
+                    $fakeFieldKey = isset($field['store_in']) ? $field['store_in'] : 'extras';
+                    $this->addCompactedField($requestInput, $fieldName, $fakeFieldKey);
 
-                if (! in_array($fakeFieldKey, $compactedFakeFields)) {
-                    $compactedFakeFields[] = $fakeFieldKey;
+                    if (! in_array($fakeFieldKey, $compactedFakeFields)) {
+                        $compactedFakeFields[] = $fakeFieldKey;
+                    }
                 }
             }
+            
         }
 
         // json_encode all fake_value columns if applicable in the database, so they can be properly stored and interpreted