+++ src/resources/views/crud/fields/date_range.blade.php
@@ -21,14 +21,14 @@
     }
 
     if (isset($entry)) {
-        $start_name = formatDate($entry, $field['start_name']);
-        $end_name = formatDate($entry, $field['end_name']);
+        $start_name = formatDate($entry, $field['name'][0]);
+        $end_name = formatDate($entry, $field['name'][1]);
     }
 ?>
 
 <div @include('crud::inc.field_wrapper_attributes') >
-    <input class="datepicker-range-start" type="hidden" name="{{ $field['start_name'] }}" value="{{ old(square_brackets_to_dots($field['start_name'])) ?? $start_name ?? $field['start_default'] ?? '' }}">
-    <input class="datepicker-range-end" type="hidden" name="{{ $field['end_name'] }}" value="{{ old(square_brackets_to_dots($field['end_name'])) ?? $end_name ?? $field['end_default'] ?? '' }}">
+    <input class="datepicker-range-start" type="hidden" name="{{ $field['name'][0] }}" value="{{ old(square_brackets_to_dots($field['name'][0])) ?? $start_name ?? $field['start_default'] ?? '' }}">
+    <input class="datepicker-range-end" type="hidden" name="{{ $field['name'][1] }}" value="{{ old(square_brackets_to_dots($field['name'][1])) ?? $end_name ?? $field['end_default'] ?? '' }}">
     <label>{!! $field['label'] !!}</label>
     <div class="input-group date">
         <input